-you can now see quizzes that were sent to you
- look at friends profile does not work yet

+++ src/app/component/show-my-quiz/show-my-quiz.ts
@@ -3,27 +3,11 @@ import {ActivatedRoute, Router} from '@angular/router';
 import { ShowMyQuizService } from './show-my-quiz-service';
 import { FormsModule } from '@angular/forms';
 import {NgForOf, NgIf} from '@angular/common';
-import {HomeService, User} from '../component/home/home.service';
-import {LoginService} from '../component/login/login.service';
+import {HomeService} from '../home/home.service';
+import {LoginService} from '../login/login.service';
+import { User } from '../../models/user-model';
+import { Quiz } from '../../models/quiz-model';
 
-interface Answer {
-  text: string;
-  correct: boolean;
-}
-
-interface Question {
-  questionName: string;
-  typeOfQuestion: string;
-  answers: Answer[];
-}
-
-interface Quiz {
-  id: number;
-  title: string;
-  description: string;
-  category: string;
-  questions: Question[];
-}
 
 @Component({
   selector: 'app-show-my-quiz',
@@ -32,10 +16,10 @@ interface Quiz {
   imports: [FormsModule, NgForOf, NgIf],
   providers: [ShowMyQuizService]
 })
+
 export class ShowMyQuiz implements OnInit {
 
   quiz?: Quiz;
-  myQuizzes: Quiz[] = [];
   currentUser!: User;
   friends: User[] = [];
   selectedFriend?: User;
@@ -64,22 +48,6 @@ export class ShowMyQuiz implements OnInit {
     });
   }
 
-  private loadMyQuiz() {
-    this.loginService.userLogin().subscribe({
-      next: (user) => {
-        this.currentUser = user;
-        //console.log('Eingeloggt als:', user);
-
-        this.homeService.getMyQuiz(user.id).subscribe({
-          next: (response) => {
-            this.myQuizzes = response;
-            //console.log('Freunde:', response);
-          }
-        });
-      }
-    });
-  }
-
   private loadQuizById() {
     const quiz_id = Number(this.route.snapshot.paramMap.get('id'));
 
@@ -92,20 +60,24 @@ export class ShowMyQuiz implements OnInit {
   }
 
   ngOnInit() {
-    const quizId = Number(this.route.snapshot.paramMap.get('id'));
+    this.loginService.userLogin().subscribe({
+      next: (user) => (this.currentUser = user)
+    });
     this.loadCurrentUserAndFriends();
-    this.loadMyQuiz();
     this.loadQuizById();
   }
 
   shareMyQuiz() {
     if (!this.quiz || !this.selectedFriend) return;
 
-    this.showMyQuizService.shareQuizWithFriend(this.quiz.id, this.selectedFriend.id)
-      .subscribe({
-        next: () => console.log(`Quiz wurde an ${this.selectedFriend?.username} geschickt`),
-        error: (err) => console.error(err)
-      });
+    if (this.currentUser.id === this.quiz.author.id) {
+      this.showMyQuizService.shareQuizWithFriend(this.quiz.id, this.selectedFriend.id)
+        .subscribe({
+          next: () => console.log(`Quiz wurde an ${this.selectedFriend?.username} geschickt`),
+          error: (err) => console.error(err)
+        });
+    }
+
   }
 
   editMyQuiz() {
@@ -115,15 +87,17 @@ export class ShowMyQuiz implements OnInit {
   deleteMyQuiz() {
     if (!this.quiz) return;
 
-    if (!confirm("Wirklich löschen?")) return;
+    if (this.currentUser.id === this.quiz.author.id) {
+      if (!confirm("Wirklich löschen?")) return;
 
-    this.showMyQuizService.deleteQuiz(this.quiz.id).subscribe({
-      next: () => {
-        console.log("Quiz gelöscht!");
-        this.router.navigate(['/home']);
-      },
-      error: (err: any) => console.error(err)
-    })
+      this.showMyQuizService.deleteQuiz(this.quiz.id).subscribe({
+        next: () => {
+          console.log("Quiz gelöscht!");
+          this.router.navigate(['/home']);
+        },
+        error: (err: any) => console.error(err)
+      })
+    }
   }
 
 }
